[CHD] Adicionando tratativas basicas no controller de Alunos

+++ app/src/main/java/application/controller/Alunos.java
@@ -3,7 +3,10 @@ package application.controller;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import application.model.Aluno;
 import application.repository.AlunoRepository;
@@ -34,7 +37,13 @@ public class Alunos {
 
     @GetMapping("/{id}")
     public Aluno details(@PathVariable long id){
-        return alunoRepo.findById(id).get();
+        Optional<Aluno> result = alunoRepo.findById(id);
+        if( result.isEmpty() ) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno não encontrado"
+            );
+        }
+        return result.get();
     }
     
     @PostMapping
@@ -42,18 +51,33 @@ public class Alunos {
         return alunoRepo.save(aluno);
     }
 
-    @PutMapping("path/{id}")
+    @PutMapping("/{id}")
     public Aluno put( 
         @PathVariable long id,    
         @RequestBody Aluno newData
     ){
         Optional<Aluno> result = alunoRepo.findById(id);
+        if( result.isEmpty() ) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno não encontrado"
+            );
+        }   
+        if( newData.getNome().isEmpty()){
+            throw new ResponseStatusException(
+                HttpStatus.BAD_REQUEST, "Nome do aluno é inválido."
+            );
+        }
         result.get().setNome(newData.getNome());
         return alunoRepo.save(result.get());     
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@RequestParam long id) {
+    public void delete(@PathVariable long id) {
+        if( !alunoRepo.existsById(id) ){
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno não encontrado"
+            );    
+        }
         alunoRepo.deleteById(id);
     }
 }
